refactor(sessions): drop the RemotePathSheet extraction the redesign made moot
The first pass at this fix asked the user for a destination path on the
remote host, which needed Profiles' private remote-path sheet lifted into
Core and parameterized on the expected extension.

Piping the export back over stdout removed that need: the save panel is now
the only destination picker and no remote path is ever entered. Put
ProfilesView back exactly as it was and drop the shared sheet, so the diff
is only the export fix.
